pyproject.toml
[project]
name = "codewiki"
version = "1.2.0"
description = "LLM-enhanced code lifecycle analysis and documentation generator"
authors = [{name = "Jay"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0", "requests>=2.28"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black>=23.0", "ruff>=0.1"]

[project.scripts]
codewiki = "codewiki.cli:main"

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

